requirements-lock.txt
# ==================== 鎖定版本依賴文件 ====================
#
# 此文件包含所有依賴的精確版本號，確保：
# ✅ 安裝速度極快（無需依賴解析）
# ✅ 環境完全可重現
# ✅ 避免版本冲突
#
# 推薦安裝方式（最快）：
#   pip install -r requirements-lock.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
#
# 如需更新依賴版本，請使用 requirements.txt 或 pyproject.toml
#
# 生成時間：2025-10-21
# Python 版本：3.10.8
# ==================== 依賴列表 ====================

aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
akshare==1.17.54
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
APScheduler==3.11.0
async-timeout==4.0.3
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
backports.asyncio.runner==1.2.0
backtrader==1.9.78.123
baostock==0.8.9
bcrypt==4.3.0
beautifulsoup4==4.13.5
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
bs4==0.0.2
build==1.3.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
chainlit==2.8.0
charset-normalizer==3.4.3
chevron==0.14.0
chromadb==1.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.1
cssselect==1.3.0
cuid==0.4
curl_cffi==0.13.0
cycler==0.12.1
dashscope==1.24.6
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
eodhd==1.0.32
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.2
feedparser==6.0.12
filelock==3.19.1
filetype==1.2.0
finnhub-python==2.4.24
flatbuffers==25.2.10
fonttools==4.60.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.8.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpath==0.82.2
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
kubernetes==33.1.0
langchain==0.3.27
langchain-anthropic==0.3.20
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-google-genai==2.1.12
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.30
Lazify==0.4.0
literalai==0.1.201
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mcp==1.14.1
mdurl==0.1.2
mini-racer==0.12.4
mmh3==5.2.0
monotonic==1.6
motor==3.7.1
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.108.2
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-alephalpha==0.47.3
opentelemetry-instrumentation-anthropic==0.47.3
opentelemetry-instrumentation-bedrock==0.47.3
opentelemetry-instrumentation-chromadb==0.47.3
opentelemetry-instrumentation-cohere==0.47.3
opentelemetry-instrumentation-crewai==0.47.3
opentelemetry-instrumentation-google-generativeai==0.47.3
opentelemetry-instrumentation-groq==0.47.3
opentelemetry-instrumentation-haystack==0.47.3
opentelemetry-instrumentation-lancedb==0.47.3
opentelemetry-instrumentation-langchain==0.47.3
opentelemetry-instrumentation-llamaindex==0.47.3
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-marqo==0.47.3
opentelemetry-instrumentation-mcp==0.47.3
opentelemetry-instrumentation-milvus==0.47.3
opentelemetry-instrumentation-mistralai==0.47.3
opentelemetry-instrumentation-ollama==0.47.3
opentelemetry-instrumentation-openai==0.47.3
opentelemetry-instrumentation-openai-agents==0.47.3
opentelemetry-instrumentation-pinecone==0.47.3
opentelemetry-instrumentation-qdrant==0.47.3
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-replicate==0.47.3
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-sagemaker==0.47.3
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation-together==0.47.3
opentelemetry-instrumentation-transformers==0.47.3
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-vertexai==0.47.3
opentelemetry-instrumentation-watsonx==0.47.3
opentelemetry-instrumentation-weaviate==0.47.3
opentelemetry-instrumentation-writer==0.47.3
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parsel==1.10.0
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
posthog==3.25.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
pypandoc==1.15
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytdx==1.72
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
questionary==2.1.1
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
stockstats==0.6.5
streamlit==1.49.1
sympy==1.14.0
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traceloop-sdk==0.47.3
# tradingagents - 本項目，需要單獨安裝：pip install -e .
tushare==1.4.24
typer==0.19.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.36.0
vine==5.1.0
w3lib==2.3.1
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
xlrd==2.0.2
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.66
zipp==3.23.0
zstandard==0.25.0
